Check if legacy ping packets are valid, instead of asuming they are
Prevents misinterpretation of legacy ping packets as handshakes.
Should fix #27

If at the begining of a connection it sees '0xFE' it assumes that it is a legacy ping packet.
Handshake packets may also start with '0xFE' though, which confuses bungee.

Bungee assumes its a ping, but doesn't do any validation at all.
Then it chokes on the partially-processed handshake packet.
This can cause all sorts of different exceptions, depending on the mood bungee is in, but they all trace back to the LegacyDecoder.

The solution is to check if the packet really is a legacy ping, and treat it normally if so.
Vanilla also does this, and encounters none of these issues.

I've tested that legacy pings still work with this commit.